+++ R/m.infos.lm.R
@@ -3,8 +3,10 @@ m.infos.lm <- function(x,
                        forminter,
                        which,
                        sig.level,
-                       aux_mt,
-                       ...) {
+                            aux_mt,
+                            ...) {
+  ci <- .tukeyc_emmeans_ci_cols(aux_mt)
+
   aux_m.inf <- aggregate(forminter,
     data = x$model,
     function(x) {
@@ -27,9 +29,8 @@ m.infos.lm <- function(x,
     "linf_se" = with(aux_mt, emmean) - abs(qt(sig.level, with(aux_mt, emmean / SE))) * aux_m.inf[[my]][, 4],
     "lsup_se" = with(aux_mt, emmean) + abs(qt(sig.level, with(aux_mt, emmean / SE))) * aux_m.inf[[my]][, 4],
     #' linf_sepool' = with(aux_mt,emmean) - abs(qt(sig.level,with(aux_mt,statistic)))*with(aux_mt,std.error),
-    "linf_sepool" = with(aux_mt, lower.CL),
-    #' lsup_sepool' = with(aux_mt,emmean) + abs(qt(sig.level,with(aux_mt,statistic)))*with(aux_mt,std.error),
-    "lsup_sepool" = with(aux_mt, upper.CL)
+    "linf_sepool" = aux_mt[[ci["lower"]]],
+    "lsup_sepool" = aux_mt[[ci["upper"]]]
   )
 
   aux_m.inf2 <- aux_m.inf1[order(aux_m.inf1[["means"]],