+++ R/m.infos.nest.aovlist.R
@@ -51,10 +51,7 @@ m.infos.nest.aovlist <- function(x,
   if (is.null(fl2)) {
     f2 <- levels(model.frame(x)[, nf1])[fl1] # corresponde ao fator onde se esta fazendo o desdobramento!
 
-    aux_m.inf21 <- subset(
-      aux_m.inf2,
-      eval(parse(text = nf1)) == f2
-    ) # pegando as medias de interesse
+    aux_m.inf21 <- .tukeyc_filter_nest(aux_m.inf2, which, fl1, fl2, model.frame(x))
 
     m.inf <- list(
       Means = aux_m.inf21[, c(1:3)],
@@ -68,10 +65,7 @@ m.infos.nest.aovlist <- function(x,
 
     f3 <- levels(model.frame(x)[, nf1])[fl1]
 
-    aux_m.inf21 <- subset(
-      aux_m.inf2,
-      eval(parse(text = nf1)) == f3 & eval(parse(text = nf2)) == f2
-    ) # pegando as medias de interesse
+    aux_m.inf21 <- .tukeyc_filter_nest(aux_m.inf2, which, fl1, fl2, model.frame(x))
 
     m.inf <- list(
       Means = aux_m.inf21[, c(1:4)],